Made front README more clear, renamed ambiguous TWIN module with appropriate Sprinkler name
