+++ Sprinkler/global_variables.py
