+++ Sprinkler/trickle.py
